+++ src/metadata/v3/codec/bz2.rs
@@ -3,7 +3,7 @@ use serde::{Deserialize, Deserializer, Serialize};
 
 /// The identifier for the `bz2` codec.
 // TODO: ZEP for bz2
-pub const IDENTIFIER: &str = "https://codec.zarrs.dev/bytes_to_bytes/bz2";
+pub const IDENTIFIER: &str = "bz2";
 
 /// A wrapper to handle various versions of `bz2` codec configuration parameters.
 #[derive(Serialize, Deserialize, Clone, Eq, PartialEq, Debug, Display, From)]