+++ src/metadata/v3/codec/pcodec.rs
@@ -3,7 +3,7 @@ use serde::{Deserialize, Deserializer, Serialize};
 
 /// The identifier for the `pcodec` codec.
 // TODO: ZEP for pcodec when stabilised
-pub const IDENTIFIER: &str = "https://codec.zarrs.dev/array_to_bytes/pcodec";
+pub const IDENTIFIER: &str = "pcodec";
 
 /// A wrapper to handle various versions of `pcodec` codec configuration parameters.
 #[derive(Serialize, Deserialize, Clone, PartialEq, Debug, Display, From)]